[micro] Add call states to state description
The caller save stack in the state description is extended with a copy
of the caller state at the time of performing a call and pushing
interesting information to said stack.

Together with newly derived equalities that compose to give us an
equality operator on those saved states, the problem of finding the
call that corresponds to a return reduces to reading the caller state
from the top of the extended stack and locating that state in the trace.