[micro] Debug failure of step_consistent
Information about the consistent failures in the "step consistency"
property. As observed, the property always fails immediately after the
execution of the first call instruction. The problem is that at that
point in the blessed call sequence, the next step inserts a new entry
into a previously unused position in the memory map. The simple
comparison on properties of memory maps before and after the step
takes place on a pair of maps with different domains (the latter
extending the former), so that simple equality is no longer adequate.