[micro] Fix simplified writeAddr, debug conj.2
First, remove the initial simplifications introduced while quickly
importing the computation of write addresses from the Forvis spec.

Second, add some more tracing to debug the computation of those
addresses. By inspecting those, we can see that there is a
disagreement between the addresses as computed by this piece of code
(and used in next_desc) and the trace of the execution reported
during testing, which appears to compute addresses differently.