[micro] Scramble after, not before calls
Adjust the scrambling scheme so that the (stack) memory is scrambled
not on execution of a call instruction (i.e., before the call), but at
the call site proper (i.e., after the call), once the accessibility
relation has changed. We leverage the extended state description to do
this implicitly at the blessed call sites instead of actively monitoring
changes in the accessibility relation -- assuming that it changes
exactly during calls (and is restored during returns).