[micro] Comment unused parts of stack testing
Af the moment, three separate attempts to capture a testing property
for stack safety coexist in the file (mine, Leo's and his refinement).
To clarify, comment out those threads that are not being used, and
consider removing them if everything in them has been superceded by
the current refined property.